requirements.txt
asgiref==3.8.1
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.4.0
cssselect==1.2.0
cssutils==2.11.1
dj-database-url==0.5.0
Django==3.2
django-extensions==2.2.9
django-grappelli==2.15.1
django-wkhtmltopdf==3.4.0
idna==3.10
lxml==5.3.0
more-itertools==10.5.0
Pillow==9.5.0
premailer==3.7.0
python-decouple==3.3
python-slugify==6.1.2
pytz==2024.2
requests==2.32.3
six==1.16.0
sqlparse==0.5.1
text-unidecode==1.3
typing-extensions==4.12.2
urllib3==2.2.3
whitenoise==5.1.0
